fix: signUpメソッドを切り分け
